Fix soft-fp fma for -Wuninitialized.
The soft-fp implementations of fma produce -Wuninitialized warnings
because, in the cases where the result is not a nonzero finite value,
the soft-fp does not set the exponent of the result since the (cooked)
packing will do so, but the compiler does not then see that the
exponent is always set in packing before it's used if it wasn't set
earlier.  This patch uses DIAG_* macros to suppress those warnings.

Tested for mips64.  (In fact this allows the mips64 build to complete
with the -Wno-uninitialized removed from math/Makefile, but more
cleanups are still needed in the ldbl-128ibm code for uninitialized
warnings there.)

	* soft-fp/fmadf4.c: Include <libc-internal.h>.
	(__fma): Ignore uninitialized warnings around packing.
	* soft-fp/fmasf4.c: Include <libc-internal.h>.
	(__fmaf): Ignore uninitialized warnings around packing.
	* soft-fp/fmatf4.c: Include <libc-internal.h>.
	(__fmal): Ignore uninitialized warnings around packing.